requirement.txt
appdirs==1.4.3
asn1crypto==0.22.0
attrs==16.3.0
Automat==0.5.0
backports.shutil-get-terminal-size==1.0.0
cffi==1.10.0
constantly==15.1.0
cryptography==1.8.1
cssselect==1.0.1
decorator==4.0.11
Django==1.9.12
django-debug-toolbar==1.7
django-extensions==1.7.8
django-mptt==0.8.7
enum34==1.1.6
idna==2.5
incremental==16.10.1
ipaddress==1.0.18
ipython==5.3.0
ipython-genutils==0.2.0
lxml==3.7.3
packaging==16.8
parsel==1.1.0
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
Pygments==2.2.0
pyOpenSSL==16.2.0
pyparsing==2.2.0
queuelib==1.4.2
scandir==1.5
Scrapy==1.3.3
selenium==3.3.3
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
Twisted==17.1.0
w3lib==1.17.0
wcwidth==0.1.7
zope.interface==4.3.3
